Some modificationr elated to memcheck

+++ src/client_context.c
@@ -1,8 +1,9 @@
 
 #include "client_context.h"
+#include <string.h>
 
 Table* lookup_table(char *name) {
-	int i;
+	size_t i;
 	for(i=0; i<(current_db->tables_size); i++){
 		if(strcmp(current_db->tables[i].name, name) == 0){
 			return &(current_db->tables[i]);
